Improve the misspelling match to handle more phonemens changes.
This approach goes from one random phoneme replacement for each
phoneme in the answer to all possible phonemes for each phoneme
in the answer.

Note: This only allows for a single phoneme to be tested at a time so
at most one phoneme change in the answer must pass the check.